feat: user-course logic

+++ database/migrations/2024_02_02_080848_create_user_courses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('course_id')->constrained()->onDelete('cascade');
-            $table->float('progress')->default(0); // Represents the percentage of course completion
+            $table->integer('progress')->default(0); // Represents the percentage of course completion
             $table->boolean('completion_status')->default(false);
             $table->timestamps();
         });
